Use an include instead of monkey patching
