--- lib/krane/cli/version.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module Krane
-  class CLI
-    desc "version", "Prints the version"
-    def version
-      logger.info("Krane Version: #{KubernetesDeploy::VERSION}")
-    end
-  end
-end
